requirements.txt
click==6.6
Flask==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysql-connector==2.1.4
PyYAML==3.12
Werkzeug==0.11.13
